Change float matrix and vector classes to double for randomized SVD

+++ RedSvd/redsvd.hpp
@@ -52,32 +52,32 @@ public:
     r = (r < A.rows()) ? r : A.rows();
     
     // Gaussian Random Matrix for A^T
-    Eigen::MatrixXf O(A.rows(), r);
+    Eigen::MatrixXd O(A.rows(), r);
     Util::sampleGaussianMat(O);
     
     // Compute Sample Matrix of A^T
-    Eigen::MatrixXf Y = A.transpose() * O;
+    Eigen::MatrixXd Y = A.transpose() * O;
     
     // Orthonormalize Y
     Util::processGramSchmidt(Y);
 
     // Range(B) = Range(A^T)
-    Eigen::MatrixXf B = A * Y;
+    Eigen::MatrixXd B = A * Y;
     
     // Gaussian Random Matrix
-    Eigen::MatrixXf P(B.cols(), r);
+    Eigen::MatrixXd P(B.cols(), r);
     Util::sampleGaussianMat(P);
     
     // Compute Sample Matrix of B
-    Eigen::MatrixXf Z = B * P;
+    Eigen::MatrixXd Z = B * P;
     
     // Orthonormalize Z
     Util::processGramSchmidt(Z);
     
     // Range(C) = Range(B)
-    Eigen::MatrixXf C = Z.transpose() * B; 
+    Eigen::MatrixXd C = Z.transpose() * B; 
     
-    Eigen::JacobiSVD<Eigen::MatrixXf> svdOfC(C, Eigen::ComputeThinU | Eigen::ComputeThinV);
+    Eigen::JacobiSVD<Eigen::MatrixXd> svdOfC(C, Eigen::ComputeThinU | Eigen::ComputeThinV);
     
     // C = USV^T
     // A = Z * U * S * V^T * Y^T()
@@ -86,22 +86,22 @@ public:
     matV_ = Y * svdOfC.matrixV();
   }
   
-  const Eigen::MatrixXf& matrixU() const {
+  const Eigen::MatrixXd& matrixU() const {
     return matU_;
   }
 
-  const Eigen::VectorXf& singularValues() const {
+  const Eigen::VectorXd& singularValues() const {
     return matS_;
   }
 
-  const Eigen::MatrixXf& matrixV() const {
+  const Eigen::MatrixXd& matrixV() const {
     return matV_;
   }
 
 private:
-  Eigen::MatrixXf matU_;
-  Eigen::VectorXf matS_;
-  Eigen::MatrixXf matV_;
+  Eigen::MatrixXd matU_;
+  Eigen::VectorXd matS_;
+  Eigen::MatrixXd matV_;
 };
 
 class RedSymEigen {
@@ -120,33 +120,33 @@ public:
     r = (r < A.rows()) ? r : A.rows();
     
     // Gaussian Random Matrix
-    Eigen::MatrixXf O(A.rows(), r);
+    Eigen::MatrixXd O(A.rows(), r);
     Util::sampleGaussianMat(O);
     
     // Compute Sample Matrix of A
-    Eigen::MatrixXf Y = A.transpose() * O;
+    Eigen::MatrixXd Y = A.transpose() * O;
     
     // Orthonormalize Y
     Util::processGramSchmidt(Y);
 
-    Eigen::MatrixXf B = Y.transpose() * A * Y;
-    Eigen::SelfAdjointEigenSolver<Eigen::MatrixXf> eigenOfB(B);
+    Eigen::MatrixXd B = Y.transpose() * A * Y;
+    Eigen::SelfAdjointEigenSolver<Eigen::MatrixXd> eigenOfB(B);
     
     eigenValues_ = eigenOfB.eigenvalues();
     eigenVectors_ = Y * eigenOfB.eigenvectors();
   }
   
-  const Eigen::MatrixXf& eigenVectors() const {
+  const Eigen::MatrixXd& eigenVectors() const {
     return eigenVectors_;
   }
 
-  const Eigen::VectorXf& eigenValues() const {
+  const Eigen::VectorXd& eigenValues() const {
     return eigenValues_;
   }
 
 private:
-  Eigen::VectorXf eigenValues_;
-  Eigen::MatrixXf eigenVectors_;
+  Eigen::VectorXd eigenValues_;
+  Eigen::MatrixXd eigenVectors_;
 };
 
 class RedPCA {
@@ -162,22 +162,22 @@ public:
   void run(const Mat& A, const int rank) {
     RedSVD redsvd;
     redsvd.run(A, rank);
-    const Eigen::VectorXf& S = redsvd.singularValues();
+    const Eigen::VectorXd& S = redsvd.singularValues();
     principalComponents_ = redsvd.matrixV();
     scores_              = redsvd.matrixU() * S.asDiagonal();
   }
 
-  const Eigen::MatrixXf& principalComponents() const {
+  const Eigen::MatrixXd& principalComponents() const {
     return principalComponents_;
   }
 
-  const Eigen::MatrixXf& scores() const {
+  const Eigen::MatrixXd& scores() const {
     return scores_;
   }
 
  private:
-  Eigen::MatrixXf principalComponents_;
-  Eigen::MatrixXf scores_;
+  Eigen::MatrixXd principalComponents_;
+  Eigen::MatrixXd scores_;
 };
 
 }
